requirements.txt
flask
pathlib
python-dotenv
PyJWT
Flask-Migrate
Flask-SQLAlchemy
flask-bcrypt
mysqlclient
